pyproject.toml
[tool.poetry]
name = "makeselfie"
version = "0.1.0"
description = "A MakeZürich 2023 project about Quantified Self"
authors = ["Oleg Lavrovsky <loleg@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7"
PySimpleGUI = "^4.60.5"
requests = "^2.31.0"
imageio-ffmpeg = "^0.4.8"

[virtualenvs.options]
system-site-packages = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
